% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2many.R
\name{rmd2many}
\alias{rmd2many}
\title{Create content in multiple formats}
\usage{
rmd2many(
  x,
  outputs = c("ioslides", "word"),
  slide_level = 1,
  css = "black",
  add18 = TRUE,
  pdf = TRUE,
  highlight = list(word = "monochrome", ioslides = NULL, slidy = NULL, html = NULL),
  ...
)
}
\arguments{
\item{x}{A numeric vector. Contains the numbers of the lectures for which
to create files. Must be a subset of `1:20`.}

\item{outputs}{A character vector. Specifies the output formats required.
A subset of `c("word", "isoslides", "slidy", "html")`.}

\item{slide_level}{Passed to [`rmd2ioslides`][accessr::rmd2ioslides] via
`...`. The default `slide_level = 1` means that a level one header #
create a new non-segue slide for an ioslides presentation.}

\item{css}{The argument \code{css} to be passed to
\code{\link[rmarkdown]{ioslides_presentation}}. If \code{css = "black"}
then \code{accessr}'s css file \code{black.css} is used, which results in
black text being used in the slides.}

\item{add18}{A logical scalar. If `TRUE` then we also create Word documents
with 18pt text.}

\item{pdf}{A logical scalar. If `TRUE` then we use
[`chrome_print`][pagedown::chrome_print] to print `lecture?.html` to
`lecture?slides.pdf`.}

\item{highlight}{A named list, with names a subset of
`c("word", "isoslides", "slidy", "html")`, providing the respective syntax
highlighting styles passed to Pandoc for the output formats. Any syntax
highlighting provided in `css` will take precedence.}

\item{...}{Additional arguments passed to
  [`rmd2ioslides`][accessr::rmd2ioslides],
  [`rmd2slidy`][accessr::rmd2slidy],
  [`rmd2word`][accessr::rmd2word] or
  [`rmd2html`][accessr::rmd2html].

  For example, the argument `params` can passed to
  \code{\link[rmarkdown]{render}}.}
}
\value{
Nothing is returned.
}
\description{
From a single R markdown file create HTML slides, PDF slides, Word
and PDF documents
}
\details{
The source `.Rmd` files must be in a directory called
  `lecture` in the current working directory.

The default setting creates, for each valid value in `x`, the following
files

  * `lecture?.html`: lecture slides is `ioslides` format.
  * `lecture?.pdf`: a PDF document containing the content of the lecture
    slides, produced from a Word file `lecture?.docx`.
  * `lecture?draft.pdf`: a version of `lecture?.pdf` in which selected
    content is removed.
  * `lecture?.docx`: a Word document containing the content of the lecture
    slides using an 18pt text template Word file. If `add18 = FALSE` then
    this is the Word file used to create `lecture?.pdf`.
}
\examples{
lecture(7:8)
}
