% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2word2pdf.R
\name{rmd2pdf}
\alias{rmd2pdf}
\alias{pns}
\alias{pnd}
\title{Converts rmarkdown code to Word and PDF documents}
\usage{
rmd2pdf(x, dir = NULL, doc = NULL, ...)

pns(x, ...)

pnd(x, ...)
}
\arguments{
\item{x}{A logical vector containing the names of the files to convert
(no extension).  May be a path and filename relative to the current
working directory, e.g., `DIRECTORY/file1`.}

\item{dir}{A apth to the directory in which the file \code{officetopdf.exe} sits.
This is not needed if this file sits in the current working directory
or a directory in the list returned by \code{searchpaths()}.
See \strong{Details}.}

\item{doc}{The name of a template Word document from which the fonts,
margins etc in the output Word document will based.  If this is not in the
working directory then also give the path.}

\item{...}{Additional arguments to be passed to \code{\link{system}}.
The argument \code{wait} determines whether or not R will wait for the
PDF files to be produced.  In the default \code{wait = TRUE} case a
warning will be given if any of the PDF files could not be produced.
This will occur if there is an existing PDF file of the same name open
in another application.  If \code{wait = FALSE} then no warnings will be
produced.}
}
\value{
A vector containing the values returned from \code{\link{system}}
  is returned invisibly. Note that if \code{wait = FALSE} then these values
  will be 0 (the success value) even if some of the PDF files could not be
  produced.  The error code 17234 indicates that a PDF file was open in
  another application.
}
\description{
Creates accessible PDF documents by creating Word documents using
rmarkdown code and then producing PDF files.
}
\details{
The \code{\link[rmarkdown]{render}} function is used to create
  a Word file in the current working directory.

  Then
  \href{https://github.com/cognidox/OfficeToPDF}{OfficeToPDF} is used to
  convert the Word file to a PDF file.  The file \code{officetopdf.exe}
  needs to be downloaded from the
  \href{https://github.com/cognidox/OfficeToPDF/releases}{OfficeToPDF releases}
  page and placed in the directory specified by the argument \code{dir}, or
  in a directory that is in the list returned by \code{\link{searchpaths}}.
  If \code{officetopdf.exe} cannot be found then an error is thrown.

  \code{pns} and \code{pnd} are convenience functions for useonly by the
  package maintainer.
}
\examples{
\dontrun{
  rmd2pdf(c("file1", "file2"), doc = "template.docx")
}
}
