% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessr-internal.R
\name{accessr-internal}
\alias{accessr-internal}
\alias{pns}
\alias{pnd}
\title{Internal accessr functions}
\usage{
pns(x, ...)

pnd(x, ...)
}
\description{
Internal accessr functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
