% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessr-internal.R
\name{accessr-internal}
\alias{accessr-internal}
\alias{rmd2presentation}
\alias{accessr_zip}
\alias{ext_img}
\title{Internal accessr functions}
\usage{
rmd2presentation(
  x,
  format = c("ioslides", "slidy"),
  zip = TRUE,
  pdf = FALSE,
  zip_pdf = zip,
  pdf_args = list(),
  add = FALSE,
  quiet = TRUE,
  rm_html = FALSE,
  rm_pdf = FALSE,
  ...
)

accessr_zip(x, dnames, udnames, zipfile, add, extension)

ext_img(src, width = 0.5, height = 0.2, alt = "", ref_docx_dim)
}
\description{
Internal accessr functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
