% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_otp.R
\name{install_otp}
\alias{install_otp}
\title{Install OfficeToPDF}
\usage{
install_otp(dir, url, method, quiet = TRUE, ...)
}
\arguments{
\item{dir}{Directory into which to download the \code{OfficeToPDF.exe}.}

\item{url}{URL from which to download \code{OfficeToPDF.exe}.  If this is
missing then
\url{https://github.com/cognidox/OfficeToPDF/releases/download/v1.9.0.2/OfficeToPDF.exe}
is used.}

\item{method, quiet}{Passed to \code{\link[utils]{download.file}}.  If
\code{method} is missing then it is inferred using
\code{getOption("download.file.method")}.}

\item{...}{Additional arguments to be passed to
\code{\link[utils]{download.file}}.}
}
\value{
See the \strong{Value} section of \code{\link[utils]{download.file}}.
}
\description{
Downloads the executable file \code{OfficeToPDF.exe} from
\url{https://github.com/cognidox/OfficeToPDF/releases}
}
\details{
\code{mode = "wb"} is used so that binary transfer is used.
}
\note{
\code{OfficeToPDF.exe} is used to create PDFs files from Word
  documents in \code{\link{rmd2word}}.
}
\seealso{
\code{\link[utils]{download.file}}, \code{\link{rmd2word}}.
}
