% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2word2pdf.R
\name{rmd2pdf}
\alias{rmd2pdf}
\title{Converts rmarkdown code to Word and PDF documents}
\usage{
rmd2pdf(x, doc, dir, zip = TRUE, add = FALSE, ...)
}
\arguments{
\item{x}{A character vector containing the names of the \code{.Rmd} files to
convert (no extension) if they are in the current working directory, or
paths to the files, either absolute or relative to the current working
directory, e.g., \code{DIRECTORY/file1}.  The \code{.docx} and \code{.pdf}
files are created in the same directory as their respective \code{.Rmd}
file.}

\item{doc}{An optional character vector to identify a template Word
document, or respective documents for the files in \code{x}, from which
the fonts, margins etc in the output Word documents will based.  If
\code{doc} is not supplied then a default is used.  Different templates
may be used for different files.  \code{rep_len(doc, length(x))} is used
to force \code{length(doc)} to have the same length as \code{x}. A
component equal to \code{"default"} may be used to use the default Word
template. See \strong{Details} for information about how to identify the
locations of the Word template files.}

\item{dir}{A path to the directory in which the file \code{officetopdf.exe}
sits.  This is not needed if this file sits in the current working
directory or a directory in the list returned by \code{searchpaths()}.
Otherwise, it may be a path relative to the current working directory
or an absolute path.}

\item{zip}{A logical scalar or character vector indicating whether the PDF
files should be put into a zip archive.  If \code{zip = FALSE} then no
zip archive is created.  Otherwise, an archive is created in each unique
directory involved in \code{x}.  If \code{zip = TRUE} then any archive
created has the name \code{accessr.zip}.  If \code{zip} is a character
vector of zip file names (no extension) then these names are used to name
the zip archives.  The names are recycled if necessary.}

\item{add}{A logical scalare that determines what happens if the output
zip file already exists.  If \code{add = TRUE} then files are added to the
zip file and if \code{add = FALSE} then the zip file is deleted and will
only contain newly-created files.}

\item{...}{Additional arguments to be passed to \code{\link{system}}.
The argument \code{wait} determines whether or not R will wait for the
PDF files to be produced.  In the default \code{wait = TRUE} case a
warning will be given if any of the PDF files could not be produced.
This will occur if there is an existing PDF file of the same name open
in another application.  If \code{wait = FALSE} then no warnings will be
produced.}
}
\value{
A vector containing the values returned from \code{\link{system}}
  is returned invisibly. Note that if \code{wait = FALSE} then these values
  will be 0 (the success value) even if some of the PDF files could not be
  produced.  The error code 17234 indicates that a PDF file was open in
  another application.
}
\description{
Creates accessible PDF documents by creating Word documents using
rmarkdown code and then producing PDF files.
}
\details{
The simplest setup is to have the \code{.Rmd} files and the Word
  template (if used) and \code{officetopdf.exe} in the current working
  directory.

  It is possible to have the \code{.Rmd} files in different
  directories, but if any non-\code{"default"} values in \code{doc} must be
  such that the \code{reference_docx} argument of
  \code{\link[rmarkdown]{word_document}} finds a template Word file.
  If the template is in the same directory as its respective \code{.Rmd}
  component in  \code{x} then the filename, e.g. \code{template.docx} will
  suffice.  Otherwise, a path to the template should be given, either
  relative to the directory in which the \code{.Rmd} file sits, or an
  absolute path.

  The \code{\link[rmarkdown]{render}} function creates a Word file from
  each input \code{.Rmd} file.  Then
  \href{https://github.com/cognidox/OfficeToPDF}{OfficeToPDF} is used to
  convert the Word file to a PDF file.  The file \code{officetopdf.exe}
  needs to be downloaded from the
  \href{https://github.com/cognidox/OfficeToPDF/releases}{OfficeToPDF releases}
  page and placed in the directory specified by the argument \code{dir}, or
  in a directory that is in the list returned by \code{\link{searchpaths}}.
  If \code{officetopdf.exe} cannot be found then an error is thrown.

  \code{pns} and \code{pnd} are convenience functions for the
  package maintainer.
}
\examples{
\dontrun{
# All files in the current working directory
rmd2pdf(c("file1", "file2"), doc = "template.docx")
}
}
